they're definitions, not constraints
